Moves enemies into their own package
