+++ src/main/java/andioopp/domain/model/tower/Towers.java
@@ -1,0 +1,11 @@
+package andioopp.domain.model.tower;
+
+import andioopp.domain.model.tower.towers.Mario;
+
+public class Towers {
+
+    public static Tower mario() {
+        return new Mario();
+    }
+
+}
