+++ src/main/java/andioopp/domain/model/tower/towers/Rosalina.java
@@ -1,0 +1,4 @@
+package andioopp.domain.model.tower.towers;
+
+public class Rosalina {
+}
