--- src/main/java/andioopp/domain/model/towers/Bobomb.java
@@ -1,4 +1,0 @@
-package andioopp.domain.model.towers;
-
-public class Bobomb {
-}
