--- src/main/java/andioopp/domain/model/towers/Luigi.java
@@ -1,4 +1,0 @@
-package andioopp.domain.model.towers;
-
-public class Luigi {
-}
