--- src/main/java/andioopp/domain/model/towers/Rosalina.java
@@ -1,4 +1,0 @@
-package andioopp.domain.model.towers;
-
-public class Rosalina {
-}
