--- src/main/java/andioopp/domain/model/towers/Toad.java
@@ -1,4 +1,0 @@
-package andioopp.domain.model.towers;
-
-public class Toad {
-}
